requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.1.1
fastapi==0.95.2
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.30.0
joblib==1.2.0
lazy_loader==0.2
llvmlite==0.40.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
onnxruntime==1.15.0
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pooch==1.7.0
protobuf==4.23.2
pydantic==1.10.8
PyMatting==1.1.8
pyreadline3==3.4.1
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.38
requests==2.31.0
scikit-image==0.20.0
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
watchdog==3.0.0
yarl==1.9.2
